Fix gaussian classifier. Change interfaces.

+++ src/main/java/ru/machine/learning/algorithms/Model.java
@@ -1,13 +1,14 @@
 package ru.machine.learning.algorithms;
 
 import io.vavr.collection.Seq;
-import tech.tablesaw.api.IntColumn;
+import tech.tablesaw.api.DoubleColumn;
 import tech.tablesaw.api.Table;
 
 import javax.annotation.Nonnull;
 
 public interface Model {
 
-    Model fit(@Nonnull Table train, @Nonnull IntColumn trainTarget);
-    Seq<Integer> predict(@Nonnull Table test);
+    Model fit(@Nonnull Table train, @Nonnull DoubleColumn trainTarget);
+
+    Seq<Double> predict(@Nonnull Table test);
 }
